Initial commit of LXPanel plugin
Finally gave in and wrote it (a couple of decades since I last touched
GTK+ in C :).  Not sure if it's useful, but at least it's.. "cute".
Once DEBs are updated, issue #2 will be complete.

Essentially, it's just a mini-frontend for DBus method calls and signals.
In it's current form, not sure if plugin will handle argonone daemon
not running (it *shouldn't* crash lxpanel and it even *might* pick up
when dbus service starts, but... dunno, tired of GLIB/GTK+ APIs right now :).