Add DEB for lxpanel plugin
This closes #2; should be ready for a new DEB release once a couple of
finishing touches to plugin are done (including making label visibility
configurable, and testing it's behavior when the D-Bus service is not up).

This commit adds an lxplug-argon1 package, which contains the lxpanel
plugin. Not sure it's done in best way, esp. since the plugin is in C and
uses a different build system (plain old makefile), so it took a bit of
debian/rules callisthenics.  However, it builds, and since I do not have
to worry about portability (this is intended only for Buster and only
on RasPi 4).. whatever.

Note that, in order to "squeeze in" the Makefile build into the
packaging process, we have to use individual override rules (which ones
to override and how was figured out mostly by trial an error, from
debhelper verbose logs).  Invoking dh twice in the "main" rule does not
work, since "dh binary" does everything from dh_auto_build up to
dh_builddeb so, if the first invocation of dh doesn't build and install
everything, then steps after dh_install will fail (and the second
invocation of dh won't even happen).  Perhaps there is a %: "delegating"
rule syntax but with a glob/pattern, but since it's just a few
overrides, doesn't matter.

Also, for some reason, a -X (exlcude) needs to be added to dh_shlibdeps,
not just to dh_makeshlibs (as in lxpanel plugins in raspberrypi-ui
github repos). Perhaps this has something to do with the fact that we
don't use autotools. However, it should be noted that "official"
lxplug-* packages for the Pi *do* seem to have redundant dependencies (to
libraries that are required by the lxpanel dependency anyway), so it's
also a possibility that the dh_shlibdeps override was forgotten there.